+++ packages/document-drive/src/server/base-server.ts
@@ -63,7 +63,6 @@ import {
 } from "document-drive";
 import { runAsap, runAsapAsync } from "document-drive/run-asap";
 import {
-  documentModelCreateDocument,
   type Action,
   type CreateDocumentActionInput,
   type DocumentModelModule,